feat: remove unnecessary json api
