+++ internal/json/api_compat.go
@@ -1,85 +1,105 @@
 //go:build !amd64 || !go1.16
 // +build !amd64 !go1.16
 
-/**
- * Copyright 2023 CloudWeGo Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2023 CloudWeGo Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 package json
 
 import (
-	"strconv"
 	"encoding/base64"
-	"encoding/json"
-
-	"github.com/cloudwego/dynamicgo/internal/native/types"
-	"github.com/cloudwego/dynamicgo/internal/rt"
+	"strconv"
+	"unicode/utf8"
+	_ "unsafe"
 )
-
-func quote(buf *[]byte, val string) {
-	*buf = append(*buf, '"')
-	quoteString(buf, val)
-	*buf = append(*buf, '"')
-}
 
 func NoQuote(buf *[]byte, val string) {
 	quoteString(buf, val)
 }
 
-func unquote(src string) (string, types.ParsingError) {
-	sp := rt.IndexChar(src, -1)
-	out, ok := unquoteBytes(rt.BytesFrom(sp, len(src)+2, len(src)+2))
-	if !ok {
-		return "", types.ERR_INVALID_ESCAPE
-	}
-	return rt.Mem2Str(out), 0
-}
+var (
+	//go:linkname safeSet encoding/json.safeSet
+	safeSet [utf8.RuneSelf]bool
 
-func decodeBase64(src string) ([]byte, error) {
-	return base64.StdEncoding.DecodeString(src)
+	//go:linkname hex encoding/json.hex
+	hex string
+)
+
+func quoteString(e *[]byte, s string) {
+	start := 0
+	for i := 0; i < len(s); {
+		if b := s[i]; b < utf8.RuneSelf {
+			if safeSet[b] {
+				i++
+				continue
+			}
+			if start < i {
+				*e = append(*e, s[start:i]...)
+			}
+			*e = append(*e, '\\')
+			switch b {
+			case '\\', '"':
+				*e = append(*e, b)
+			case '\n':
+				*e = append(*e, 'n')
+			case '\r':
+				*e = append(*e, 'r')
+			case '\t':
+				*e = append(*e, 't')
+			default:
+				// This encodes bytes < 0x20 except for \t, \n and \r.
+				// If escapeHTML is set, it also escapes <, >, and &
+				// because they can lead to security holes when
+				// user-controlled strings are rendered into JSON
+				// and served to some browsers.
+				*e = append(*e, `u00`...)
+				*e = append(*e, hex[b>>4])
+				*e = append(*e, hex[b&0xF])
+			}
+			i++
+			start = i
+			continue
+		}
+		c, size := utf8.DecodeRuneInString(s[i:])
+		// if c == utf8.RuneError && size == 1 {
+		//     if start < i {
+		//         e.Write(s[start:i])
+		//     }
+		//     e.WriteString(`\ufffd`)
+		//     i += size
+		//     start = i
+		//     continue
+		// }
+		if c == '\u2028' || c == '\u2029' {
+			if start < i {
+				*e = append(*e, s[start:i]...)
+			}
+			*e = append(*e, `\u202`...)
+			*e = append(*e, hex[c&0xF])
+			i += size
+			start = i
+			continue
+		}
+		i += size
+	}
+	if start < len(s) {
+		*e = append(*e, s[start:]...)
+	}
 }
 
 func encodeBase64(src []byte) string {
 	return base64.StdEncoding.EncodeToString(src)
-}
-
-func (self *Parser) decodeValue() (val types.JsonState) {
-	e, v := DecodeValue(self.s, self.p)
-	if e < 0 {
-		return v
-	}
-	self.p = e
-	return v
-}
-
-func (self *Parser) skip() (int, types.ParsingError) {
-	e, s := SkipValue(self.s, self.p)
-	if e < 0 {
-		return self.p, types.ParsingError(-e)
-	}
-	self.p = e
-	return s, 0
-}
-
-func (self *Node) encodeInterface(buf *[]byte) error {
-	out, err := json.Marshal(self.packAny())
-	if err != nil {
-		return err
-	}
-	*buf = append(*buf, out...)
-	return nil
 }
 
 func i64toa(buf *[]byte, v int64) {
