+++ internal/json/error.go
@@ -1,45 +1,25 @@
-/**
- * Copyright 2023 CloudWeGo Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2023 CloudWeGo Authors.
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 package json
 
 import (
 	"fmt"
 	"strings"
-	"unsafe"
 
 	"github.com/cloudwego/dynamicgo/internal/native/types"
 )
-
-func (self *Parser) syntaxError(err types.ParsingError) SyntaxError {
-	return SyntaxError{
-		Pos:  self.p,
-		Src:  self.s,
-		Code: err,
-	}
-}
-
-func newSyntaxError(err SyntaxError) *Node {
-	msg := err.Description()
-	return &Node{
-		t: V_ERROR,
-		v: int64(err.Code),
-		p: unsafe.Pointer(&msg),
-	}
-}
 
 type SyntaxError struct {
 	Pos  int
